Safe mode added + more logging + admin safeguards

+++ app/media/routes.py
@@ -4,7 +4,7 @@ from app import db
 from datetime import datetime
 from dateutil.relativedelta import relativedelta
 from flask_login import login_required, current_user
-from flask import render_template, request
+from flask import render_template, request, current_app
 from app.scripts.media import check_pick_creation, delete_media_everywhere, check_user_creation, check_movie_creation, check_tv_show_creation
 from app.models import Media, Movie, TVShow, Pick, AppSettings
 
@@ -75,9 +75,15 @@ def add_pick(media_id):
 @login_required
 def delete_pick(pick_id):
     pick = Pick.query.get_or_404(pick_id)
+    current_app.logger.info("User "+current_user.alias+" is trying to delete pick "+str(pick))
+    if (not current_user.admin) and (not current_user==pick.user):
+        return {
+            "error" : "Not allowed!"
+        }, 405
     media = pick.media
     db.session.delete(pick)
     db.session.commit()
+    current_app.logger.info("User "+current_user.alias+" deleted pick "+str(pick))
     check_if_abandonned(media)
     return{
         "message" : "Pick of "+media.title+" deleted"
@@ -86,7 +92,12 @@ def delete_pick(pick_id):
 @bp.route("/<int:media_id>/delete", methods=['DELETE'])
 @login_required
 def delete_media(media_id):
+    if not current_user.admin:
+        return {
+            "error" : "Not allowed!"
+        }, 405
     media = Media.query.get_or_404(media_id)
+    current_app.logger.info("User "+current_user+" is trying to delete media "+ media)
     script_result = delete_media_everywhere(media)
     return {
         "message" : script_result
